# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Teste Back-end PHP - Drugovich(V1)",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/grupos": {
            "get": {
                "tags": [
                    "Grupos"
                ],
                "summary": "Listar grupos",
                "description": "Retornar todos os grupos cadastrados",
                "operationId": "listar_grupos",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário ou você não tem permissão! ",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno!",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Grupos"
                ],
                "summary": "Cadastrar novo grupo",
                "operationId": "novo_grupo",
                "requestBody": {
                    "description": "Campos para cadastros",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "nome",
                                        "format": "text",
                                        "example": "Marcos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário ou você não tem permissão! "
                    },
                    "417": {
                        "description": "Entradas inválidas, campos obrigatórios! "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/grupos/{id}": {
            "get": {
                "tags": [
                    "Grupos"
                ],
                "summary": "Detalhar informação do grupo  e clientes vinculados",
                "description": "Retornar informações do grupo e clientes vinculados",
                "operationId": "get_grupo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário ou você não tem permissão! ",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "500": {
                        "description": "Erro interno!",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Grupos"
                ],
                "summary": "Alterar cadastro do grupo",
                "operationId": "editar_grupo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Campos para cadastros",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "nome",
                                        "format": "text",
                                        "example": "Marcos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Não autorizado, login é necessário ou você não tem permissão! "
                    },
                    "417": {
                        "description": "Entradas inválidas, campos obrigatórios! "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Grupos"
                ],
                "summary": "Excluir cadastro do grupo",
                "operationId": "excluir_grupo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tudo Certo!",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Não autorizado, login é necessário ou você não tem permissão! "
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearerAuth"
            }
        }
    ]
}